Remove unneeded arguments
The staking indicator tool tip for no mature coins was erroneously given these arguments while debugging the stake weight. In my rush to debug the weight, I did the same (these arguments) in a proper location and forgot to remove the original debugging attempt code (these arguments) after the debugging was complete and when I did remove the correct debugging code. Sorry for my oversight. It was the result of sleeplessness and not carelessness.