Fixed initial download blocked every 500 PoS block
from:
Peershares/commit/ad71a3b1890208c9ef8ea005c76ee18fbf0b6065
The problem is Proof of Stake blocks are not stored when the previous
transaction of the CoinStake is missing. So the initial download process
is broken.

Peercoin solved this by using the latest Proof of Work block as
reference to download, but it doesn't work for Peershares where we only
have PoS blocks.

To solve this:
The reference block is now the best block, whatever the proof.

The client asks for missing blocks when it receives a PoS block and
it doesn't have the CoinStake previous transaction.

Block request are retried immediatly during initial download because
the same last block is asked multiple times until we have the full
chain.